refactor(memory): split vector doctor orchestration
